backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
sentence-transformers==2.3.1
pinecone-client==3.0.0
requests==2.31.0
numpy==1.26.3
torch==2.1.2
